DMI: add support for retrieving SMBIOS entry point via UEFI
In UEFI-based systems, the SMBIOS entry point can be retrieved from
the UEFI configuration tables. With this commit, the UEFI
bootloader retrieves the SMBIOS entry point and passes it to the
kernel via a newly defined region type (REGION_SMBIOS); the kernel
uses the SMBIOS entry point, if known, to find the location of the
DMI tables.
This allows retrieving the DMI tables in UEFI-based systems, which
is needed by the cloud-init klib to properly detect the cloud
platform on which the kernel is running.