Move fs path code into tfs and add fuse write callbacks In order to avoid duplicated code in the fuse program, this change consolidates system-independent fs object/path lookup and resolution code that was scattered in filesystem.c and unix.c and moves it into tfs.c. One compromise is that the path lookup code needs to know about the root filesystem as well as have the ability to cross mount points which requires system-specific code, so these two things have been moved into functions which must be initialized by system code before using the filesystem.
This change also implements most of the write callbacks for fuse so the
fuse driver is can read and modify a disk image now.

Finally, since the fuse program requires the fuse kernel driver and libraries,
it was moved to a separate new ADDITIONAL_PROGRAMS list in the makefile
so that it must be built separately and won't cause the build to fail for anyone
who does not have or does not want to have the fuse driver installed.