Radar: add more robust retry mechanisms
The radar klib now checks the server response to crash reports, and
retries sending a crash report if the response is not received or
does not contain an OK status code.
A retry mechanism is implemented for boot reports as well, so that
a boot report is resent if no valid boot ID has been received.
Finally, a retry is triggered if connection to the server fails
when sending either a crash or a boot report.