User address space: replace id heaps with rangemap-based heap
Id heaps use an underlying bitmap for storing information on what
pages are allocated (1 bit for each page). This doesn't scale when
using 4KB pages if the allocated address space reaches a few
terabytes.
This change replaces the virtual and virtual_page id heaps of the
user process with a rangemap-based heap. An address range
allocated on the virtual heap but not mapped is represented by a
struct vmap whose flags have the VMAP_FLAG_MMAP bit cleared; as a
result, it cannot be mapped on-demand when a page fault occurs
(unless the user process requests a mapping later on). Memory
areas being mapped with VMAP_FLAG_PREALLOC are not released upon
unmapping, so that the underlying address range is not released in
the virtual heap until an explicit deallocation is requested.