Aarch64: fix user virtual address limit and page mapping code
Aarch64 uses a 4-level page table set for mapping virtual
addresses, where each level covers 512 table entries; given a 4KB
page size, this sets the maximum number of bits to be used for a
virtual address range to (12 + 9 + 9 + 9 + 9) = 48. As a result,
user virtual addresses must be limited to the (0 - ffffffffffff)
range.

In addition, this commit fixes the calculation of the end address
when remapping an already mapped range, so that the range passed to
the next page table level does not exceed the maximum length that
the next level can handle.