EPOLL_CTL_ADD: fix setting of event mask and data
When a file descriptor is added to an epoll file, the corresponding
epollfd structure must be initialized with the user-supplied event
mask and data.
The existing code was failing to do the above in cases where the
file descriptor being added was already present in the fd vector of
the epoll file. This change fixes the issue by defining a new
function called reset_epollfd() that properly resets an
already-allocated epollfd structure with new values for the event
mask and data; this function is now used for both
epoll_ctl(EPOLL_CTL_ADD) and poll().