Page cache: remove page refcount increment in touch_or_fill_page()
This function is called in many different scenarios, including
whenever an active page is being read from the cache. Incrementing
the page refcount at each such occurrence means that reading a page
from the cache makes it non-evictable.
This commit fixes the above issue by removing the calls to
refcount_reserve() from touch_or_fill_page(), and replacing them
with calls in the different parts of the code where bumping the
refcount is actually needed.