extents_range_handler: properly handle extend failure when allocating (#1481)
If extents_range_handler was called during the allocation stage of a tfs write
and a subsequent call to update_extent_length returned a failure, an
apply_merge would be called with a null pointer. This changes the extend call
to mirror fill_gap, returning a failure status instead of trying to apply it
directly to a merge. Also, since the presence of a merge indicates extent
writing is in process, eliminate the redundant write flag parameter in
extent_range_handler, fill_gap and extend.