add flush method to storage interfaces with implementations for virtio_storage and virtio_scsi (#1485)
This extends the block device interface to add a "flush" method. A flush is
issued as a final step in the chain of operations initiated by
filesystem_flush, used by the various paths that synchronize storage
devices. Flush implementations are included for virtio-storage (which issues a
VIRTIO_BLK_T_FLUSH command) and virtio-scsi (which issues a SYNCHRONIZE_CACHE
SCSI command) drivers; other storage drivers will require flush
implementations to be added later.

As part of the virtio-storage implementation, the VIRTIO_BLK_F_FLUSH and
VIRTIO_BLK_F_CONFIG_WCE features are negotiated. A flush method is created and
writeback cache mode is set depending on negotiation of respective
features. This improves the virtio-storage performance drastically, which now
out-performs the virtio-scsi driver in the bulk write test by over 10% in one
Linux/KVM-based test setup. This speedup is to be expected, as noted in the
OASIS virtio spec, because the virtio-blk device involves much less overhead
than the virtio-scsi device.

(Note that the overall performance for the write test will appear to drop
slightly when compared to that of previous commits. This is because the
storage flush operation is now included as part of the measured duration.)