syscall.c: rename_internal(): avoid referencing current thread
It is not necessary to call set_syscall_return() or
set_syscall_error() in syscall-specific code that doesn't run in a
bottom half context: simply returning the relevant exit code is
enough.
This change removes all calls to set_syscall_return() and
set_syscall_error() from rename_internal(), so that file rename
operations can be done in contexts other than a syscall execution,
where there may not be a valid current thread.