Add syslog klib
This klib implements a console driver, which can be configured via
the manifest to output console messages to a file (with a simple
log rotation algorithm) and/or to an UDP syslog server.
The `syslog` manifest tuple is used to configure syslog behavior.
Supported configuration options:
- file: if present, syslog output is written to a file (whose path
is indicated by the value of this option)
- file_max_size: file size at which log rotation occurs (default
value: 8 MB); its value contains a byte count, optionally followed
by a multiplier suffix ('k' or 'K' for KB, 'm' or 'M' for MB, 'g
or 'G' for GB)
- file_rotate: number of rotated files to keep in the filesystem
(default value: 1); if 0, old logs are deleted instead of being
rotated
- server: if present, syslog output is sent to an UDP server (whose
hostname or IP address is indicated by the value of this option)
- server port: UDP port number (default value: 514) at which syslog
output is sent
At least one of the `file` or `server` options must be specified.
Example manifest contents for file output:
```
syslog:(
  file:var/log/syslog
)
```
Example manifest contents for UDP output:
```
syslog:(
  server:example.com
)
```