add "cwd" manifest option for setting current working directory (#1500)
The current working directory for an application may now be set by including a
path string under the "cwd" attribute in the program manifest. A test for this
feature is added to the mkdir runtime test. To allow parity testing with
Linux, the mkdir test has also been updated to use a test directory under
/tmp (rather than the root directory) and create files with legitimate
permissions.