klibs: implement automatic loading with dependency resolution
With this change, all files found in the "klib" folder of the klib
filesystem are automatically loaded when the kernel starts. If a
given klib depends on another klib that has not been initialized,
its init function returns KLIB_MISSING_DEP, the kernel inserts the
klib in a retry list, and the init function is called again after
other klibs have been initialized. If one or more dependencies
cannot be resolved, or if an automatically loaded klib init
function returns any other error code, the kernel halts.
The radar klib has been changed to use the dependency resolution
implemented in the kernel instead of manually loading the tls klib.
To do this, it calls the newly defined get_klib_sym() function,
which looks up a given symbol among all symbols exported by all
initialized klibs; this function allows klibs to check whether a
given symbol has been exported without requiring callers to have a
reference to the klib that is supposed to export the symbol.

The klib_handler closure has been changed to take as argument the
return value of a klib init function instead of a status code, so
that the handler of automatically loaded klibs can properly detect
missing dependencies between klibs.

The test klib has been moved to a "test" subfolder, so that it is
not automatically loaded and its specific klib handler can be
invoked instead.