Program start: wait until klibs have been initialized
This eliminates the need for user programs that depend on klibs to
wait for klibs to be initialized.
The init_klib() function now takes a status handler as additional
parameter, and calls the status handler when all automatically
loaded klibs have been initialized.
The code that handles volume mounting has been modified to call a
status handler instead of a thunk upon completion, so that
storage_when_ready() can be passed the status handler returned by
apply_merge().