Klibs: tls_close(): nullify application layer closures before calls
There may be edge cases where a TLS connection failure or
connection shutdown triggers a fatal error for the kernel, which
leads to a halt() call that does not return. To properly handle
such cases, tls_close() should set the stored application layer
closure pointers to 0 before invoking the closures, so that if
another network packet is received afterwards there will not be
further invocations of these closures.