Klibs: tls: fix handling of connections closed by the peer
When the remote peer of a TLS connection signals that the
connection is going to be closed, mbedtls_ssl_read() returns
MBEDTLS_ERR_SSL_PEER_CLOSE_NOTIFY. This condition is not
necessarily an error, and thus should not cause the incoming data
handler to return an error status code. Analogously, a shutdown of
the underlying TCP connection by the remote peer should not cause a
status error code to be returned.
These changes allow handling gracefully use cases where the kernel
downloads content from a remote server and the end of the data
being downloaded is signaled by closing the connection.