HTTP parser: fix detection of the end of HTTP header lines
The end of a header line in an HTTP message is always signaled by
the CRLF sequence, thus the internal state of the HTTP parser
should be switched to STATE_HEADER (or STATE_BODY) only when the LF
character is encountered.
This change fixes a bug where the last LF character of the HTTP
headers is mistankenly considered as the first byte of the HTTP
body.