Klibs: cloud_init: add support for file downloads
The cloud_init klib can now be configured via a "cloud_init"
manifest tuple to download one or more files from remote servers
before starting the user program.
The syntax of the manifest tuple is as below:
```
cloud_init:(
  download:[
    (src:<url> dest:<file>)
    ...
  ]
)
```
`<url>` is the URL from which a given file neds to be downloaded;
only URL values beginning with either `http://` or `https://` are
supported.
<file> is the filesystem path where a downloaded file is saved.
The `download` attribute value can consist of multiple tuples, each
of which indicates a file to be downloaded.
If a given file is already present in the filesystem and its length
is non-zero, it is not downloaded when the kernel boots, unless the
`overwrite` attribute is specified in the download directive, as in
the example below:
```
(src:http://example.com/file1 dest:file1 overwrite:t)
```
Only direct download links are supported (no redirects).
Chunked transfer ecoding is NOT supported.
If one or more download links begin with `https://`, the cloud_init
klib requires the tls klib as a dependency.