Release script: add optional smoke tests on cloud providers
The release.sh script now takes an optional command line argument,
which if present specifies a command to be executed before
publishing the release artifacts.
Currently, the only supported command is `smoke-test`, which
triggers execution of smoke tests on different cloud providers
(currently, gcp and aws). A smoke test consists of creating a
cloud image with the webg program, starting an instance from this
image, sending 1000 HTTP requests to this instance, and checking
that it responds correctly; the created instance and image are then
deleted as a final cleanup step. If any of these steps fails, the
release artifacts are not published.
The location of config.json files for Ops can be specified via
environment variables:
- OPS_CONF_GCP specifies the location of the config file for gcp
- OPS_CONF_AWS specifies the location of the config file for aws
- OPS_CONF specifies the location of a generic config file, which
is used for a given cloud provider if no provider-specific
environment variable is set.
Example:
```
export OPS_CONF_GCP=~/config-gcp.json
export OPS_CONF_AWS=~/config-aws.json
./release.sh smoke-test
```
These changes are compatible with Linux and macOS shells.
Closes #1015.