enable interrupts while spinning for kernel lock
This change addresses deadlock situations that can occur when acquiring the
kernel lock on multi-processor configurations. Previously, a call to
kern_lock (the only such call occurs in syscall_schedule) would cause a cpu to
spin with interrupts disabled. In one scenario, an in-kernel page fault
results in a page mapping, with completion of the fault contingent on
acknowledgement of flush IPIs from all other running processors. The processor
spin waiting on the kernel lock would then be unable to process the flush IPI,
and thus the kernel page fault would never complete, resulting in a
deadlock. In another scenario, the sole cpu on which storage interrupts are
handled could spin wait on the kernel lock while a kernel fault on a
file-backed page is in progress. Similar to the flush IPI scenario, the
storage interrupt would not be handled, preventing completion of the page
fault and resulting in a deadlocked kernel.

The more general solution to this problem is to enable interrupt handling by
default during kernel processing. Until widespread interrupt safety is
accomplished, this change enables interrupt handling within kern_lock,
accompanied by a slight modification to common_handler (irq_handler for
aarch64) that returns to an interrupted kernel context when in cpu_kernel
state (as opposed to calling runloop if in cpu_idle or cpu_user states). This
same logic can be used to allow interrupt handling during kernel execution in
general.