spin_rlock(): check for write lock before incrementing reader count
In a heavy lock contention scenario with one writer and many
readers trying to acquire a read/write spinlock (which can happen
in VMs with a high CPU count), there can be a deadlock where the
reader count in the lock never becomes zero, which prevents the
writer from fully acquiring the lock.
This change fixes the above issue by modifying spin_rlock() so that
the reader count is not incremented if a writer has (at least
partially) acquired the lock.