SMP: add support for an arbitrary number of vCPUs
This change removes the MAX_CPUS constant and replaces the cpuinfos
static array with a dynamically allocated vector.
The 64-bit idle CPU mask in the scheduler is replaced by a bitmap,
which is initialized in the newly defined function
init_scheduler_cpus(); since the idle CPU mask is accessed by
interrupt handlers and runloop code, init_scheduler_cpus() is
called before probing any devices (such as disks and network
interfaces) which could generate interrupts. The call to
init_secondary_cores(), where the total number of CPUs is
determined, is being moved to just before the call to
init_scheduler_cpus().
init_cpuinfo() is being modified in order to allocate dynamically
what previously was in a static array.
The fixed-size cpu_set_t type, used in the thread affinity mask
and the sched_getaffinity() and sched_setaffinity() syscalls, is
being removed, and the above syscalls now have a u64 * parameter
instead, while the thread affinity mask is now a bitmap.
In x86, the per-CPU GDT and TSS memory areas are now part of the
struct cpuinfo_machine instead of being statically defined in
assembly code, and are initialized in the init_cpuinfo_machine()
function.