Tools: dump: restore terminal foreground color to default value
When displaying filesystem contents as a tree (with the `-t`
command line option), the dump tool was changing the terminal
foreground color to white, which could make the terminal less
readable, or even unusable (if the foreground color is also white).
This change fixes the above issue by restoring the terminal
foreground color to its default value (instead fo white) after
printing a colored filesystem entry.