buffer_extend(): fix handling of consumed buffers
When less than `len` bytes are available after the end of the
current contents, buffer_extend() calculates empirically a new
capacity value based on the length of the current contents and the
`len` value.
If the buffer has been at least partially consumed from, the
calculated capacity value may be less than or equal to the current
capacity. If it is equal, then buffer_set_capacity() does not do
anything, which leaves the buffer in a state not suitable for
appending `len` bytes; if it is less, then the buffer memory area
is reallocated with a smaller size, which is not necessary.
This commit fixes the above issues by modifying buffer_extend() so
that buffer_set_capacity() is called only if the current capacity
value has to be increased, otherwise the current buffer contents
are moved to the beginning of the allocated memory to make room for
appending new data.
The buffer unit tests have been enhanced with a new test case that
would have failed without this fix.