page cache: fix zeroing of newly allocated pages
When the end of data being written to a file is not page-aligned,
the memory from the end of the data to the end of the last page
should all be zeroed (not just until the end of the last block),
because if the file size is later extended without writing data to
the last page, any data from the last page can be read when reading
the file.
With this commit, the code zeroes all the data until the end of the
page. A test case that would have failed without this fix is also
being added to the write runtime test.