TFS: fix handling of files not linked to the filesystem
When a file is not linked to the filesystem (e.g. as a result of
an unlink or rename syscall), its metadata and extent tuples can
potentially be destroyed at any time when the TFS log is compacted.
Thus, they should not be accessed, even if the file may still
exist due to open file descriptors.
This commit modifies the code that handles tuples associated to a
file so that it accesses these tuples only if the `md` member of
struct fsfile is non-zero (which indicates that the file is linked
to the filesystem). filesystem_alloc() and filesystem_dealloc() now
take an fsfile argument instead of a tuple, because they can be
called for files without an associated metadata tuple.
filesystem_creat_unnamed() no longer creates a tuple (because files
opened with O_TMPFILE are not linked to the filesystem), and
returns an fsfile instead of a tuple.
filesystem_rename() is now fixed so that if a file is being deleted
(because it's being overwritten by another file), its reference
count is decreased instead of deallocating its extents, because the
file could still be referenced by open file descriptors.
When a file is finally deallocated (i.e. when its reference count
becomes zero), deallocate_fsfile() is now called instead of
file_extents_dealloc() (which was leaking the extent rangemap).
deallocate_fsfile() has been modified to be able to use a generic
extent destructor closure.
The fallocate and write runtime tests have been amended to exercise
parts of the code that handles files not linked to the filesystem.