interrupt handlers: move thread enqueue to follow interrupt processing
The external interrupt dispatch routines were re-enqueueing the interrupted
user thread before setting the FRAME_FULL flag and calling interrupt
handlers. This could cause a "frame already full" fault to occur if such a
thread were to be migrated to another CPU and executed before the FRAME_FULL
set occurred in interrupt dispatch. The migrated thread would run and take an
interrupt or issue a syscall, at which point the FRAME_FULL flag would have
been set by the original interrupt dispatch, causing a terminal fault.

This fix simply moves the call to schedule_frame to follow setting of the
FRAME_FULL flag. Even if the thread is migrated to another CPU prior to
runloop service on the originating CPU, the set flag will be cleared upon
return to the user frame.