Interim update to make gdb stub work for user programs
This change updates the in-kernel gdb stub to work again with multi-threaded
user programs running on multiple cpus. It breaks out the architecture
specific parts to allow for future ARM debugging support. This adds support
for thread commands, status, and id. The gdb exception handler was updated
to pass through certain exceptions like page faults, and the handler is
now installed for all frames of all threads on the user process.

Breakpoint handling has been updated allowing hardware breakpoints
to be installed on all cpus, and using the correct code of hardware
breakpoints. This allows the use of software breakpoints which are much
less limiting for line-by-line debugging. Software breaks are not
implemented directly, but deferred to gdb so it will manually insert
trap instructions. Text pages are force updated with write permission
to allow traps to be inserted, although it's not a long term solution.

A new queue is added to the cpuinfo struct to allow thunks to be run
for that specific cpu. This is necessary for hardware breakpoints to
be modified on every cpu. This should also be temporary until thunks
can be scheduled per cpu.

There is further work to be done on the stub so this should be
considered a work-in-progress update with expected limits and gaps
in stub functionality.