epoll: fix edge trigger behavior
When epoll_wait() is called, the current event status is checked
for all registered file descriptors by calling check_fdesc(), which
notifies the current events to the epoll_wait_notify() handler.
The events notified to the handler should reflect the current
status of a file descriptor, regardless of the last events set on
the epoll instance for that descriptor, because the handler uses
the notified events to update (set and/or clear as appropriate) the
last events. The existing code was removing the last events from
the events reported to the handler, which causes the last events to
toggle between successive invocations of epoll_wait() even if the
status of the file descriptor does not change.
A new test case (test_edgetrigger() function) has been added to the
epoll runtime tests to trigger the above issue: without this fix,
the third call to epoll_wait() would report the EPOLLOUT event for
both the first and the third file descriptor, even though the
status of the first file descriptor does not change between the
first and the third call to epoll_wait().
In order to avoid breaking the peculiar edge trigger behavior of
eventfd, the eventfd edge handler has been modified so that both
EPOLLIN and EPOLLOUT are reported after a read or write, to mirror
the Linux implementation. This behavior is now tested in the epoll
runtime tests.