Network sockets: fix syscall handling for shut down TCP sockets
When both TX and RX directions of a TCP socket are shut down, the
`info.tcp.lw` netsock struct member is set to NULL, because the
lwIP PCB structure can be deallocated at any time and should not be
referenced anymore. If other syscalls are invoked on a shut down
socket, they should deal properly with this state and avoid
accessing a NULL pointer.
A few more test cases have been added to the netsock runtime tests
to exercise handling of sockets after shutdown.