accept(): return socket file descriptor even if connection is reset
The current code makes the accept() syscall return -ECONNRESET if
the TCP socket just created has been disconnected already. This
error code is not among the errors returned by other operating
systems for this syscall (even Linux, which may return network
errors on the new socket as error codes from accept(), never
returns -ECONNRESET), and is likely to be considered as a fatal
error by user programs (notably, it causes Golang web server
functions such as http.ListenAndServe() and
http.ListenAndServeTLS() to exit with error).
This commit changes accept_bh() so that it returns the file
descriptor of the newly created socket even if the socket is
disconnected (in this case, the peer address cannot be retrieved
and will be set to a dummy value). As a side effect, the newly
created socket is no longer leaked, because it can be closed by the
user program via the returned file descriptor, and the EPOLLIN flag
is correctly reset on the listening socket if there are no more
incoming connections.