poll.c: remove unnecessary calls to resolve_fd_noret()
The epollfd struct now has an fdesc member field; if this field is
set when a file descriptor is registered, there is no need to call
resolve_fd_noret() whenever the fdesc structure corresponding to a
given file descriptor needs to be accessed. This member field can
be used if and only if a file descriptor is registered (if the
fdesc structure associated to a registered file descriptor is
deallocated, the file descriptor is automatically unregistered by
the notify handler).
The epoll runtime tests now check that the closure of a file
descriptor registered in an epoll instance is handled gracefully by
the instance.