Unix file descriptors: implement proper reference counting
Whenever an fdesc structure is retrieved from the list of open file
descriptors, its reference count needs to be bumped until the
structure is no longer used. This is to avoid accessing invalid
memory if a file descriptor is closed while the underlying fdesc
structure is still used (e.g. in the middle of a read or write).
This commit changes all file descriptor accesses to use fdesc_get()
and fdesc_put(), which update the fdesc reference count
appropriately. For all socket callbacks (e.g. sendto, recvfrom,
etc.), the onus of calling fdesc_put() is on each implementation of
a callback (because it may block and thus not return to the
caller).
A couple of error return codes from syscalls have been fixed in the
process, in order to match Linux kernel behavior. Some test cases
have been added to the runtime tests to exercise parts of the code
that has been changed.
Even though updating the fdesc reference count is currently
necessary only for syscalls that block, it is being done also in
syscalls that don't block, because this will become necessary when
the global kernel lock is removed.