exit_thread(): remove unnecessary call to blockq_flush_thread()
A thread can only be terminated by itself, thus exit_thread() can
only be called on a thread in a runnable state.
This commit removes the unnecessary call to blockq_flush_thread()
(if this call was necessary, the current code would be unsafe
without the global kernel lock, because the blocked_on thread
structure member could be changed by another core while the
exit_thread function is accessing it).