Thread struct: protect against concurrent access
This change adds a spinlock to the thread structure, and uses this
spinlock to protect structure members against concurrent access.
Also, thread_from_tid() now increments the thread reference count,
to prevent a thread from being deallocated while it is being
accessed; callers of thread_from_tid() need to call
thread_release() when they are done with the thread.