Process struct: protect against concurrent access
This change adds a spinlock to the process structure, and uses this
spinlock to protect structure members against concurrent access.
The code that needs to retrieve the current working directory and
its filesystem now calls process_get_cwd(), which retrieves this
information atomically with respect to the chdir and fchdir
syscalls.
For x86_64, the virtual32 heap has been switched from non-locking
to locking.