Revise blockq with locking and other fixes
To support syscall execution outside of the domain of the kernel lock, the
blockq code is revised to allow operations on a blockq to occur safely on
multiple processors. This essentially amounts to protecting the list of
waiting threads per blockq with a spinlock. Other improvements that have been
made include the embedding of the previous "blockq_item" struct directly into
the thread struct, which obviates the need to (de)allocate items on each
block, adding blockq_reserve() and blockq_release() calls to support safe
implementation of thread_attempt_interrupt(), and associating any
io_completions for a given blocking operation with each waiting thread instead
of with the blockq, thus allowing multiple waiting threads to have their own
io_completions, rather than only a single I/O completion for a given
resource (blockq).

The comments at the head of blockq.c were very outdated and have been updated
to reflect the current behavior and usage.