TFS: refactor access to filesystem metadata for volume mounting
This commit removes direct access to filesystem metadata from
kernel/storage.c and unix/filesytem.c (which will be unsafe when
the kernel lock is removed), by adding the filesystem_mount()
function and replacing the lookup_follow() fs_path_helper callback
with a new get_mountpoint() callback. The code in storage.c sets
this callback to storage_get_mountpoint(), which has been modified
to return (beside the mount point) the filesystem that hosts the
mount point.