Filesystem syscalls: remove use of resolve_cstring()
Currently, most syscalls that modify the filesystem and take a file
path as argument resolve the path to a pair of tuples (one for the
file and one for the parent directory), do a series of checks on
the resolved tuples, and finally call a TFS function that does the
requested filesystem modification. Similarly, most syscalls that
read from the filesystem resolve the supplied path to a tuple, and
then retrieve the requested information from the tuple.
When the kernel lock is removed, it will be unsafe to reference
tuples resolved via resolve_cstring() or resolve_cstring_follow(),
because those tuples can potentially be unlinked from the
filesystem by another thread and then deallocated at any time.
This commit changes the above syscalls so that file path resolution
and handling of filesystem tuples is done inside the TFS code
(where protection of a filesystem from concurrent accesses can be
more easily implemented); in cases where reading of filesystem
metadata is best done at the syscall level, the new
filesystem_get_node() function is now called: this function
resolves the supplied path to a tuple (possibly creating a new file
in the filesystem) and returns that tuple (as well as an fsfile
pointer, if applicable), which can be safely accessed from external
code until filesystem_put_node() is called.
Between filesystem_get_node() and filesystem_put_node(), no other
filesystem functions can be called, because the filesystem may be
locked (to allow safe access to metadata).
Actual locking of the filesytem wil be added in a future commit.
The error code returned by truncate() when the supplied path does
not correspond to either a file or a directory has been changed
from ENOENT to EINVAL, to match Linux kernel behavior.
A few checks have been added to some runtime tests.
The FUSE driver code has been adapted to use the modified
filesystem functions.