use spinlocks to protect entry points into lwIP, socket state changes and consumption from socket incoming queues
These changes begin the process of making networking code safe to run without
protection of the kernel lock. Entry points into lwIP and socket state changes
are protected by a global spinlock, and consumption of data from socket
incoming queues is protected with a per-socket spinlock. For all socket blockq
actions, the lwIP lock is only taken if the action is invoked from a syscall
top half (not called with BLOCKQ_ACTION_BLOCKED). Blocked invocations (that
are not also nullifications) are exclusively a result of a wakeup from an lwIP
callback, where the lwIP lock would already be held.

Additionally, this resolves a potential race in connect_tcp_bh by closing any
connection established before a nullifying invocation could be handled.