pagecache_allocate_node(): remove unnecessary closure allocations
The key_compare and print_key closures passed to init_rbtree()
don't need to be allocated for each page cache node, since no
node-specific information needs to be stored in them. Instead, they
can be embedded in the pagecache structure. This will remove the
need for deallocating these structures in
pagecache_deallocate_node(), so that this function can be
implemented without having to access rbtree structure members
directly.